Script to enter gdb.
Because of the way the build process is performed, we must use
libtool to access the executable. Because of this, to debug the
program, we use the command
	libtool --mode=execute gdb main
making sure that we are in the appropriate directory.
It might also be useful to be able to pass arguments through to gdb.